requirements.txt
opencv-python
pytesseract
numpy
python-dotenv
regex
